requirements.txt
Flask
flask_sqlalchemy
psycopg2-binary
pdfplumber
python-docx
spacy
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
Werkzeug
